View definition fixed

+++ scores/views.py
@@ -11,13 +11,13 @@ from django.http import Http404
 
 @api_view(['GET','POST'])
 def list_images(request, format=None):
-    def get(self):
+    if request.method == 'GET':
+        print('HERE')
         images = Features.objects.all()
         serializer = FeaturesSerializer(images, many=True)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
-
-    def post(self, request, format=None):
+    elif request.method == 'POST':
         serializer = FeaturesSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -25,32 +25,32 @@ def list_images(request, format=None):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
-def image_features(request, format=None):
+def image_features(request, file_name,format=None):
     """
     Retrieve, update or delete a snippet instance.
     """
-    def get_object(self, pk):
+    def get_object(pk):
         try:
-            return Features.objects.get(pk=pk)
+            return Features.objects.get(file_name=str(pk))
         except Features.DoesNotExist:
             raise Http404
 
-    def get(self, request, pk, format=None):
-        features = self.get_object(pk)
+    if request.method=='GET':
+        features = get_object(file_name)
         serializer = FeaturesSerializer(features)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
     
 
-    def put(self, request, pk, format=None):
-        features = self.get_object(pk)
+    elif request.method == 'PUT':
+        features = get_object(pk)
         serializer = FeaturesSerializer(features, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk, format=None):
-        snippet = self.get_object(pk)
+    elif request.method == 'DELETE':
+        snippet = get_object(pk)
         snippet.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -65,7 +65,7 @@ def image_score(request, format=None):
         except Features.DoesNotExist:
             raise Http404
 
-    def get(self, request, pk, format=None):
+    if request.method == 'GET':
         features = self.get_object(pk)
         serializer = ScoreSeriaizer(features)
         return Response(serializer.data)
